majore structural refactor
- created `Slack.Client` namespace
- refactored all Slack classes into individual `.cs`
- refactored, renamed & moved `SlackClient` class, removing
`PostWebhookMessage` method
- added `Authentication` class to `SlackClient`, supporting reading API
token from an XML file
- added `SlackRequest` class scaffold
- cleaned up namespaces, removing any traces of `DynaSlack`
- added `Slack.Helpers` namespace and addded `Web` class to it, which is
a refactor of old `Helpers.cs`
- moved Slack responses into `SlackResponse` class, from previous
`Helpers.cs`
