+++ src/DynaSlack/DynaSlack.csproj
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+<Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
     <ResolveAssemblyWarnOrErrorOnTargetArchitectureMismatch>
       None
@@ -32,7 +32,7 @@
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>DynaSlack</RootNamespace>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
@@ -109,9 +109,17 @@
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
       <DependentUpon>Settings.settings</DependentUpon>
     </Compile>
-    <Compile Include="Web\Helpers.cs" />
-    <Compile Include="Slack\SlackClasses.cs" />
-    <Compile Include="Slack\SlackClient.cs" />
+    <Compile Include="Slack\Client\Authentication.cs" />
+    <Compile Include="Slack\Requests\SlackPayload.cs" />
+    <Compile Include="Slack\Classes\Message.cs" />
+    <Compile Include="Slack\Classes\Bot.cs" />
+    <Compile Include="Slack\Classes\Attachment.cs" />
+    <Compile Include="Slack\Classes\Webhook.cs" />
+    <Compile Include="Slack\Requests\SlackRequest.cs" />
+    <Compile Include="Slack\Requests\SlackResponse.cs" />
+    <Compile Include="Helpers\Web.cs" />
+    <Compile Include="Slack\Classes\Field.cs" />
+    <Compile Include="Slack\Client\SlackClient.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -144,8 +152,7 @@
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
   </ItemGroup>
-  <ItemGroup>
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="AfterClean">
     <!-- Remove build folders -->
