+++ src/DynaSlack/Properties/AssemblyInfo.cs
@@ -33,7 +33,7 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("2017.1.3.6")]
-[assembly: AssemblyFileVersion("2017.1.3.6")]
+[assembly: AssemblyVersion("2017.2.0.0")]
+[assembly: AssemblyFileVersion("2017.2.0.0")]
 [assembly: NeutralResourcesLanguage("en-GB")]
