+++ src/DynaSlack/keys.xml
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<Slack>
+  <token>Bearer 0/330c026b8db6e3a3a0d5f4d5907bbe6f</token>
+</Slack>
