+++ src/DynaSlack/packages.config
@@ -2,5 +2,5 @@
 <packages>
   <package id="DynamoVisualProgramming.DynamoServices" version="1.2.0" targetFramework="net45" />
   <package id="Newtonsoft.Json" version="8.0.3" targetFramework="net45" />
-  <package id="RestSharp" version="105.2.3" targetFramework="net45" />
+  <package id="RestSharp" version="105.2.3" targetFramework="net45" requireReinstallation="true" />
 </packages>
