+++ src/DynaSlack/pkg.json
@@ -2,7 +2,7 @@
 "license": "AGPL-3.0",
 "file_hash": null,
 "name": "DynaSlack",
-"version": "2017.1.3",
+"version": "2017.2.0",
 "description": "Slack Integration for Dynamo",
 "group": "",
 "keywords": [
